requirements.txt
pyngrok>=5.1.0
PyYAML>=6.0
requests>=2.27.1
gmail-connector>=0.5.0
pyOpenSSL>=22.0.0
python-dotenv>=0.19.2